logger usage

+++ engine/include/Fizziks/Log.h
@@ -40,5 +40,5 @@ FIZZIKS_API void clearLogSinks();
 }
 
 #ifndef ACTIVE_LOG_LEVEL
-#define ACTIVE_LOG_LEVEL Fizziks::LogLevel::ERROR
+#define ACTIVE_LOG_LEVEL static_cast<int>(Fizziks::LogLevel::ERROR)
 #endif
